#35 re organized project structure to separate enclave module from sample enclave
