--- protocolAggregator/aggregator-commons/build.gradle
@@ -1,18 +1,0 @@
-apply plugin: 'kotlin'
-apply plugin: 'org.jetbrains.dokka'
-
-dokka {
-    outputFormat = 'gfm'
-    outputDirectory = "$rootDir/github-docs"
-}
-
-dependencies {
-    implementation "com.r3.conclave:conclave-common:$conclaveVersion"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    implementation "net.i2p.crypto:eddsa:0.3.0"
-
-    implementation 'org.jetbrains:annotations:19.0.0'
-    testImplementation "org.junit.jupiter:junit-jupiter:5.6.0"
-    implementation 'org.apache.commons:commons-lang3:3.5'
-    implementation 'org.apache.avro:avro:1.10.2'
-}
