+++ protocolAggregator/workflows/src/test/kotlin/com/protocol180/aggregator/flow/CoalitionConfigurationTest.kt
@@ -68,9 +68,9 @@ class CoalitionConfigurationTest {
                 provider2.info.chooseIdentityAndCert().party.name))
 
         val dataTypes = listOf(CoalitionDataType("testDataType1", "Test Data Type 1",
-                ClassLoader.getSystemClassLoader().getResourceAsStream("testSchema1.avsc").readFully(), "com.protocol180.aggregator.clientEnclave.ExampleAggregationEnclave"),
+                ClassLoader.getSystemClassLoader().getResourceAsStream("testSchema1.avsc").readFully(), "com.protocol180.aggregator.sample.ExampleAggregationEnclave"),
             CoalitionDataType("testDataType2", "Test Data Type 2",
-                ClassLoader.getSystemClassLoader().getResourceAsStream("testSchema2.avsc").readFully(), "com.protocol180.aggregator.clientEnclave.ExampleAggregationEnclave"))
+                ClassLoader.getSystemClassLoader().getResourceAsStream("testSchema2.avsc").readFully(), "com.protocol180.aggregator.sample.ExampleAggregationEnclave"))
 
         val flow1 = CoalitionConfigurationUpdateFlow(coalitionPartyToRole, dataTypes)
         val future1 = host.startFlow(flow1)
